+++ CHANGELOG.md
@@ -6,6 +6,9 @@
 
 - fix wasm physics initialization when using buffered evaluation mode
 
+- clamp `MmdStandardMaterial.alpha` value to 0.0 .. 1.0 for better result
+    - you can disable this behavior by setting `MmdStandardMaterial.clampAlpha` to `false`
+
 ## 0.47.1 (2024-06-14)
 
 - package.json update for npm publish